+++ src/main/java/fi/dy/masa/malilib/gui/GuiColorEditorHSV.java
@@ -477,28 +477,28 @@ public class GuiColorEditorHSV extends GuiDialogBase
         int cw = this.sizeHS;
         int ch = 16;
 
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // H
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // H
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // S
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // S
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // V
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // V
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // R
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // R
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // G
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // G
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // B
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // B
         y += yd;
-        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF, z); // A
+        RenderUtils.drawOutline(x, y, w, h, 0xC0FFFFFF); // A
 
         x = this.xHS;
         y = this.yHS;
         w = this.sizeHS;
         h = this.sizeHS;
 
-        RenderUtils.drawOutline(x - 1, y - 1, w + 2, h + 2, 0xC0FFFFFF, z); // main color selector
-        RenderUtils.drawOutline(cx - 1, cy - 1, cw + 2, ch + 2, 0xC0FFFFFF, z); // current color indicator
-        RenderUtils.drawOutline(this.xHFullSV, y - 1, this.widthHFullSV, this.sizeHS + 2, 0xC0FFFFFF, z); // Hue vertical/full value
+        RenderUtils.drawOutline(x - 1, y - 1, w + 2, h + 2, 0xC0FFFFFF); // main color selector
+        RenderUtils.drawOutline(cx - 1, cy - 1, cw + 2, ch + 2, 0xC0FFFFFF); // current color indicator
+        RenderUtils.drawOutline(this.xHFullSV, y - 1, this.widthHFullSV, this.sizeHS + 2, 0xC0FFFFFF); // Hue vertical/full value
 
         Tessellator tessellator = Tessellator.getInstance();
         BufferBuilder buffer = tessellator.begin(VertexFormat.DrawMode.QUADS, VertexFormats.POSITION_TEXTURE);