+++ src/main/java/fi/dy/masa/malilib/mixin/input/MixinMouse.java
@@ -1,4 +1,4 @@
-package fi.dy.masa.malilib.mixin;
+package fi.dy.masa.malilib.mixin.input;
 
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.Mouse;