+++ src/main/java/fi/dy/masa/malilib/mixin/network/MixinClientConfigurationNetworkHandler.java
@@ -1,11 +1,10 @@
-package fi.dy.masa.malilib.mixin;
+package fi.dy.masa.malilib.mixin.network;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 import net.minecraft.client.network.ClientConfigurationNetworkHandler;
 import net.minecraft.network.packet.s2c.config.ReadyS2CPacket;
 import net.minecraft.registry.DynamicRegistryManager;