+++ src/main/java/fi/dy/masa/malilib/mixin/render/IMixinDrawContext.java
@@ -1,4 +1,4 @@
-package fi.dy.masa.malilib.mixin;
+package fi.dy.masa.malilib.mixin.render;
 
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.render.VertexConsumerProvider;