+++ src/main/java/fi/dy/masa/malilib/mixin/render/MixinDrawContext.java
@@ -1,4 +1,4 @@
-package fi.dy.masa.malilib.mixin;
+package fi.dy.masa.malilib.mixin.render;
 
 import java.util.List;