+++ src/main/java/fi/dy/masa/malilib/util/Color4f.java
@@ -3,11 +3,12 @@ package fi.dy.masa.malilib.util;
 import java.util.Objects;
 
 /**
- * New Version will be located under util/data
+ * New Version will be located under util/data {@link fi.dy.masa.malilib.util.data.Color4f}
  */
 @Deprecated
 public class Color4f
 {
+    public static final Color4f WHITE = new Color4f(1.0F, 1.0F, 1.0F, 1.0F);
     public static final Color4f ZERO = new Color4f(0F, 0F, 0F, 0F);
     public final float r;
     public final float g;