+++ src/main/java/fi/dy/masa/malilib/util/position/BlockRotation.java
@@ -1,27 +1,35 @@
 package fi.dy.masa.malilib.util.position;
 
-import org.jetbrains.annotations.ApiStatus;
+import java.util.function.IntFunction;
+import io.netty.buffer.ByteBuf;
 
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.codec.PacketCodecs;
+import net.minecraft.util.StringIdentifiable;
+import net.minecraft.util.function.ValueLists;
 import net.minecraft.util.math.Direction;
 
 import fi.dy.masa.malilib.MaLiLibReference;
+import fi.dy.masa.malilib.config.IConfigOptionListEntry;
 import fi.dy.masa.malilib.util.StringUtils;
 
 /**
  * Post-ReWrite code
  */
-@ApiStatus.Experimental
-public enum BlockRotation
+public enum BlockRotation implements IConfigOptionListEntry, StringIdentifiable
 {
-    NONE    (0, net.minecraft.util.BlockRotation.NONE, "none"),
-    CW_90   (1, net.minecraft.util.BlockRotation.CLOCKWISE_90, "rotate_90"),
-    CW_180  (2, net.minecraft.util.BlockRotation.CLOCKWISE_180, "rotate_180"),
-    CCW_90  (3, net.minecraft.util.BlockRotation.COUNTERCLOCKWISE_90, "rotate_270");
+    NONE    (0, net.minecraft.util.BlockRotation.NONE,                  "none"),
+    CW_90   (1, net.minecraft.util.BlockRotation.CLOCKWISE_90,          "rotate_90"),
+    CW_180  (2, net.minecraft.util.BlockRotation.CLOCKWISE_180,         "rotate_180"),
+    CCW_90  (3, net.minecraft.util.BlockRotation.COUNTERCLOCKWISE_90,   "rotate_270");
 
+    public static final EnumCodec<BlockRotation> CODEC = StringIdentifiable.createCodec(BlockRotation::values);
+    public static final IntFunction<BlockRotation> INDEX_TO_VALUE = ValueLists.createIdToValueFunction(BlockRotation::getIndex, values(), ValueLists.OutOfBoundsHandling.WRAP);
+    public static final PacketCodec<ByteBuf, BlockRotation> PACKET_CODEC = PacketCodecs.indexed(INDEX_TO_VALUE, BlockRotation::getIndex);
     public static final BlockRotation[] VALUES = values();
 
     private final int index;
-    private final String name;
+    private final String configString;
     private final String translationKey;
     private final net.minecraft.util.BlockRotation vanillaRotation;
 
@@ -29,18 +37,30 @@ public enum BlockRotation
     {
         this.index = index;
         this.vanillaRotation = vanillaRotation;
-        this.name = name;
-        this.translationKey = MaLiLibReference.MOD_ID + ".label.block_rotation." + name;
+        this.configString = name;
+        this.translationKey = MaLiLibReference.ID + ".label.block_rotation." + name;
     }
 
-    public String getName()
+    public int getIndex()
     {
-        return this.name;
+        return this.index;
+    }
+
+    @Override
+    public String getStringValue()
+    {
+        return this.configString;
     }
 
     public String getDisplayName()
     {
         return StringUtils.translate(this.translationKey);
+    }
+
+    @Override
+    public String asString()
+    {
+        return this.configString;
     }
 
     public BlockRotation add(BlockRotation rotation)
@@ -83,6 +103,12 @@ public enum BlockRotation
         return VALUES[index];
     }
 
+    @Override
+    public IConfigOptionListEntry fromString(String value)
+    {
+        return byName(value);
+    }
+
     public net.minecraft.util.BlockRotation getVanillaRotation()
     {
         return this.vanillaRotation;
@@ -92,7 +118,7 @@ public enum BlockRotation
     {
         for (BlockRotation rot : VALUES)
         {
-            if (rot.name.equalsIgnoreCase(name))
+            if (rot.configString.equalsIgnoreCase(name))
             {
                 return rot;
             }