+++ src/main/java/org/thinkingstudio/mafglib/MaFgLib.java
@@ -21,7 +21,7 @@ public class MaFgLib {
         if (FMLLoader.getDist().isClient()) {
             modContainer.registerExtensionPoint(ConfigScreenEntrypoint.class, new ModMenuImpl());
             MaLiLib.onInitialize();
-            MaLiLibDataGen.onInitializeDataGenerator(modEventBus);
+            modEventBus.addListener(MaLiLibDataGen::onInitializeDataGenerator);
         }
     }
 }