+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Changelog
-- remove `ModInitializer`
+- sync `sakura-ryoko/malilib` 1.21.4-0.23.3
+- tag namespace `mafglib` -> `malilib`
+- update NeoForge version to `21.4.136`